requirements.txt
# AI & Computer Vision (Raspberry Pi)
ultralytics>=8.0.0          # YOLOv8-nano for fruit detection
opencv-python>=4.8.0        # Image preprocessing
tflite-runtime>=2.14.0      # TensorFlow Lite for MobileNetV2
numpy>=1.24.0               # Numerical operations
pillow>=10.0.0              # Image handling

# Raspberry Pi Hardware Control
picamera2>=0.3.0            # Camera interface for Raspberry Pi
RPi.GPIO>=0.7.1             # GPIO control for hardware
gpiozero>=2.0.0             # Simplified GPIO interface

# Web Interface
flask>=3.0.0                # Web framework
flask-cors>=4.0.0           # CORS support
flask-socketio>=5.3.0       # Real-time communication
python-socketio>=5.10.0     # SocketIO client
eventlet>=0.33.0            # Async server

# Utilities
pyyaml>=6.0                 # Configuration files
python-dotenv>=1.0.0        # Environment variables
loguru>=0.7.0               # Enhanced logging

# ============================================
# Training Dependencies (PC/Laptop with GPU)
# Install these ONLY on training machine, NOT on Raspberry Pi
# ============================================
# torch>=2.0.0                # PyTorch for YOLOv8 training
# torchvision>=0.15.0         # Vision utilities
# tensorflow>=2.14.0          # TensorFlow for MobileNetV2
# matplotlib>=3.7.0           # Visualization
# scikit-learn>=1.3.0         # Metrics and evaluation
# labelImg                    # YOLO annotation tool
